회원가입시 로그인 이미지 고정. 카테고리, 검색 -> 채분 시 지역코드 int로 넘기기.

+++ app/src/main/java/com/E2I3/chaebunchaebun/HomeFragment.java
@@ -68,7 +68,7 @@ public class HomeFragment extends Fragment {
     String[] address = {"",};
     String userId = null;
     String nickname = "";
-    long locationCode = 0;
+    int locationCode = 0;
     int isNew = 0;
 
     public HomeFragment() {
@@ -195,7 +195,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 0);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -213,7 +213,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 1);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -231,7 +231,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 2);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -249,7 +249,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 3);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -267,7 +267,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 4);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -285,7 +285,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 5);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -303,7 +303,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 6);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -321,7 +321,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 7);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -339,7 +339,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 8);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
 
@@ -357,7 +357,7 @@ public class HomeFragment extends Fragment {
                 Bundle categoryBundle = new Bundle();
                 categoryBundle.putInt("vegetable", 9);
                 categoryBundle.putString("userId", userId);
-                categoryBundle.putLong("locationCode", locationCode);
+                categoryBundle.putInt("locationCode", locationCode);
                 FragmentTransaction categoryTransaction = getActivity().getSupportFragmentManager().beginTransaction();
                 categoryTransaction.setCustomAnimations(R.anim.anim_slide_in_right, R.anim.anim_slide_out_left, R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
